+++ backend/canisters/user/impl/src/governance_clients/mod.rs
@@ -1,2 +1,0 @@
-pub mod nns;
-pub mod sns;